refactor: rename collection from bitbucket to ops
