+++ .github/workflows/release.yml
@@ -6,7 +6,7 @@ on: # yamllint disable-line rule:truthy
 jobs:
   releaseanddeploy:
     runs-on: ubuntu-latest
-    if: github.repository == 'i2btech/ansible-collection-bitbucket'
+    if: github.repository == 'i2btech/ansible-collection'
     strategy:
       matrix:
         python-version: [3.10.11]
@@ -42,7 +42,7 @@ jobs:
           tag_name: v${{ steps.cversion.outputs.version }}
           release_name: Release v${{ steps.cversion.outputs.version }}
           body: |
-            # Ansible Collection: i2btech.bitbucket
+            # Ansible Collection: i2btech.ops
 
             Install with:
             ```sh
@@ -53,7 +53,7 @@ jobs:
             ```yaml
             ---
             collections:
-                - https://github.com/i2btech/ansible-collection-bitbucket/releases/download/v${{ steps.cversion.outputs.version }}/i2btech-bitbucket-${{ steps.cversion.outputs.version }}.tar.gz
+                - https://github.com/i2btech/ansible-collection/releases/download/v${{ steps.cversion.outputs.version }}/i2btech-ops-${{ steps.cversion.outputs.version }}.tar.gz
             ```
           draft: false
           prerelease: false
@@ -65,9 +65,9 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         with:
           upload_url: ${{ steps.create_release.outputs.upload_url }}
-          asset_path: i2btech-bitbucket-${{ steps.cversion.outputs.version }}.tar.gz
-          asset_name: i2btech-bitbucket-${{ steps.cversion.outputs.version }}.tar.gz
+          asset_path: i2btech-ops-${{ steps.cversion.outputs.version }}.tar.gz
+          asset_name: i2btech-ops-${{ steps.cversion.outputs.version }}.tar.gz
           asset_content_type: application/tar+gzip
 
       - name: Deploy Ansible collection to Galaxy
-        run: ansible-galaxy collection publish i2btech-bitbucket-${{ steps.cversion.outputs.version }}.tar.gz --api-key ${{ secrets.ANSIBLE_GALAXY_API_KEY }}
+        run: ansible-galaxy collection publish i2btech-ops-${{ steps.cversion.outputs.version }}.tar.gz --api-key ${{ secrets.ANSIBLE_GALAXY_API_KEY }}
