+++ src/galaxy.yml
@@ -5,7 +5,7 @@
 namespace: i2btech
 
 # The name of the collection. Has the same character restrictions as 'namespace'
-name: bitbucket
+name: ops
 
 # The version of the collection. Must be compatible with semantic versioning
 version: 1.0.5
@@ -21,7 +21,7 @@ authors:
 
 ### OPTIONAL but strongly recommended
 # A short summary description of the collection
-description: Modules to manage Bitbucket Cloud
+description: Modules to manage operational services of I2B (Bitbucket Cloud, Google Workspace, etc)
 
 # Either a single license or a list of licenses for content inside of a collection. Ansible Galaxy currently only
 # accepts L(SPDX,https://spdx.org/licenses/) licenses. This key is mutually exclusive with 'license_file'
@@ -43,16 +43,16 @@ tags: []
 dependencies: {}
 
 # The URL of the originating SCM repository
-repository: https://github.com/i2btech/ansible-collection-bitbucket
+repository: https://github.com/i2btech/ansible-collection
 
 # The URL to any online docs
-documentation: https://github.com/i2btech/ansible-collection-bitbucket/blob/master/README.md
+documentation: https://github.com/i2btech/ansible-collection/blob/master/README.md
 
 # The URL to the homepage of the collection/project
-homepage: https://github.com/i2btech/ansible-collection-bitbucket
+homepage: https://github.com/i2btech/ansible-collection
 
 # The URL to the collection issue tracker
-issues: https://github.com/i2btech/ansible-collection-bitbucket/issues
+issues: https://github.com/i2btech/ansible-collection/issues
 
 # A list of file glob-like patterns used to filter any files or directories that should not be included in the build
 # artifact. A pattern is matched from the relative path of the file or directory of the collection directory. This
