+++ src/plugins/modules/repo.py
@@ -48,7 +48,7 @@ author:
 
 EXAMPLES = r'''
 - name: "Create example repository"
-  i2btech.bitbucket.repo:
+  i2btech.ops.repo:
   username: "alice"
   password: "app_password"
   repository: "example-X"
@@ -65,7 +65,7 @@ message:
 '''
 
 #pylint: disable=wrong-import-position
-from ansible_collections.i2btech.bitbucket.plugins.module_utils.bitbucket import BitbucketHelper
+from ansible_collections.i2btech.ops.plugins.module_utils.bitbucket import BitbucketHelper
 from ansible.module_utils.basic import AnsibleModule
 #pylint: disable=wrong-import-position
