+++ src/plugins/modules/repo_perm.py
@@ -60,7 +60,7 @@ author:
 
 EXAMPLES = r'''
 - name: "Set permissions for repository X"
-  i2btech.bitbucket.repo_perm:
+  i2btech.ops.repo_perm:
   username: "alice"
   password: "app_password"
   repository: "example-X"
@@ -82,7 +82,7 @@ message:
 '''
 
 #pylint: disable=wrong-import-position
-from ansible_collections.i2btech.bitbucket.plugins.module_utils.bitbucket import BitbucketHelper
+from ansible_collections.i2btech.ops.plugins.module_utils.bitbucket import BitbucketHelper
 from ansible.module_utils.basic import AnsibleModule
 #pylint: disable=wrong-import-position
