+++ src/tests/sample-pb-repo-env.yml
@@ -6,7 +6,7 @@
   tasks:
 
     - name: Test repo_env module
-      i2btech.bitbucket.repo_env:
+      i2btech.ops.repo_env:
         username: "xxx"
         password: "xxx"
         repository: "poc-sample"
@@ -20,7 +20,7 @@
             secured: True
 
     - name: Test repo_env module
-      i2btech.bitbucket.repo_env:
+      i2btech.ops.repo_env:
         username: "xxx"
         password: "xxx"
         repository: "poc-sample"
